types added so that it is clear what one gets

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/controller/AnswerController.java
@@ -19,7 +19,7 @@ public class AnswerController {
     @ResponseStatus(HttpStatus.CREATED)
     @ResponseBody
     public void saveAnswers(@PathVariable("gamePin") int gamePin,
-                            @PathVariable("roundNumber") Long roundNumber,
+                            @PathVariable("roundNumber") int roundNumber,
                             @RequestHeader("Authorization") String userToken,
                             @RequestBody Map<String, String> answers) {
